feat: enhance ProfileActivity layout and user data display
